pyproject.toml
[project]
name = "gmail-mcp-server"
version = "0.1.0"
description = "MCP server for Gmail integration"
authors = [
    {name = "AJ Brown", email = "aj@example.com"},
]
dependencies = [
    "mcp @ file:///Users/ajbrown/Desktop/claude/python-sdk",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
gmail-mcp-server = "gmail_mcp_server.server:main"

[tool.hatch.metadata]
allow-direct-references = true